requirements_local.txt
requests==2.28.1
pandas==2.2.2
numpy==2.2.5
python-telegram-bot[job_queue]==20.0
matplotlib==3.6.2
python-dotenv==1.0.1